+++ packages/helpers/jest.config.js
@@ -1,0 +1,27 @@
+let mappedModule;
+switch (process.env.TEST_ENV) {
+  case "cjs":
+    mappedModule = "<rootDir>/dist/curi-helpers.js";
+    break;
+  case "umd":
+    mappedModule = "<rootDir>/dist/curi-helpers.umd.js";
+    break;
+  default:
+    mappedModule = "<rootDir>/src/index";
+}
+
+module.exports = {
+  preset: "ts-jest",
+  moduleFileExtensions: ["ts", "js"],
+  testMatch: ["**/tests/**/*.spec.ts"],
+  globals: {
+    "ts-jest": {
+      module: "es6",
+      diagnostics: false
+    }
+  },
+  testURL: "http://localhost",
+  moduleNameMapper: {
+    "@curi/helpers": mappedModule
+  }
+};