+++ packages/helpers/types/index.d.ts
@@ -1,0 +1,3 @@
+import once from "./once";
+import preferDefault from "./preferDefault";
+export { once, preferDefault };