+++ packages/router/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./types";
 
 import curi from "./curi";
-import once from "./utils/once";
 import pathname from "./interactions/pathname";
 
-export { curi, once, pathname };
+export { curi, pathname };