+++ packages/router/types/index.d.ts
@@ -1,5 +1,4 @@
 export * from "./types";
 import curi from "./curi";
-import once from "./utils/once";
 import pathname from "./interactions/pathname";
-export { curi, once, pathname };
+export { curi, pathname };