+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Remove the `once()` export (moved to `@curi/helpers`).
+
 ## 1.0.3
 
 * Add dev guard to log to only call in development.